Update to use shared and makeStyles
