+++ app/ui/src/components/AddCard.tsx
@@ -1,14 +1,33 @@
-import { Card, Text } from '@fluentui/react-components'
+import { Card, makeStyles, Text, tokens } from '@fluentui/react-components'
 import { AddRegular } from '@fluentui/react-icons'
-import useStyles from '../styles/useStyles'
+
 
 interface AddCardProps {
   onClick: () => void
   labelText: string
 }
 
+const componentSyles = makeStyles({
+  addIconText: {
+    textAlign: 'center',
+    fontWeight: tokens.fontWeightSemibold
+  },
+  disabledCard: {
+    width: '200px',
+    maxWidth: '100%',
+    height: '260px'
+  },
+  addIcon: {
+    fontSize: '72px',
+    color: tokens.colorNeutralForeground3,
+    margin: 'auto',
+    opacity: 0.5,
+    pointerEvents: 'none'
+  }
+})
+
 const AddCard: React.FC<AddCardProps> = ({ onClick, labelText }) => {
-  const classes: Record<string, string> = useStyles()
+  const componentClasses = componentSyles()
 
   const handleClick = onClick || (() => {})
 
@@ -16,12 +35,12 @@ const AddCard: React.FC<AddCardProps> = ({ onClick, labelText }) => {
     <Card
       aria-label={labelText}
       role="button"
-      className={classes.disabledCard}
+      className={componentClasses.disabledCard}
       appearance="filled-alternative"
       onClick={handleClick}
     >
-      <AddRegular className={classes.addIcon} />
-      <Text className={classes.addIconText}>{labelText}</Text>
+      <AddRegular className={componentClasses.addIcon} />
+      <Text className={componentClasses.addIconText}>{labelText}</Text>
     </Card>
   )
 }
