+++ app/ui/src/components/PageHeader.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import aiDocIcon from '../assets/ai-doc.png'
-import useStyles from '../styles/useStyles'
+import { sharedStyles } from '../styles/sharedStyles'
+import { makeStyles } from '@fluentui/react-components'
+
 
 interface PageHeaderProps {
   title: string
@@ -8,13 +10,26 @@ interface PageHeaderProps {
   customElement?: React.ReactNode
 }
 
+const componentStyles = makeStyles({
+  headerContent: {
+    margin: '16px',
+    color: '#000',
+    maxWidth: '800px',
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+})
+
 const PageHeader: React.FC<PageHeaderProps> = ({ title, description, customElement }) => {
-  const classes = useStyles()
+  const sharedClasses = sharedStyles()
+  const componentClasses = componentStyles()
 
   return (
-    <div className={classes.container}>
-      <div className={classes.headerContent}>
-        <div className={classes.heroImage}>
+    <div className={sharedClasses.container}>
+      <div className={componentClasses.headerContent}>
+        <div className={sharedClasses.heroImage}>
           <img src={aiDocIcon} alt="AI Document Review Icon" width="150" />
         </div>
         <div>
