+++ app/ui/src/pages/admin/AgentManager.tsx
@@ -1,23 +1,24 @@
 import { Divider } from '@fluentui/react-components';
 import PageHeader from '../../components/PageHeader';
 import AgentList from '../../components/AgentList';
-import useStyles from '../../styles/useStyles';
+import { sharedStyles } from '../../styles/sharedStyles';
+
 
 function AgentManager() {
-  const classes = useStyles();
+  const sharedClasses = sharedStyles();
 
   return (
-    <div className={classes.container}>
+    <div className={sharedClasses.container}>
       {/* Header Section */}
       <PageHeader
         title="Manage Agents"
         description="View and update the AI doc review agents by updating existing agent prompts or by adding
             new ones to enhance your document analysis."
       />
-      <Divider className={classes.divider} />
+      <Divider className={sharedClasses.divider} />
 
       {/* Agents List and Add Agent */}
-      <div className={classes.agentsContainer}>
+      <div className={sharedClasses.agentsContainer}>
         <AgentList/>
       </div>
     </div>
